// src/enums/EntityMessages.ts
export enum EntityMessages {
    //Send an external acceleration to this object.
    ACCELERATE = 'accelerate',
    TRY_ATTACK = 'tryattack',
    TRY_DASH = 'trydash',
    LAUNCH_ATTACK = 'launchattack',
    HIT_BY_ATTACK = 'hitbyattack',
    HIT_BLOCK = 'hitblock',
    CHANGE_HP = 'changehp',
    CHANGE_HP_MAX = 'changehp',
    BATTLE_WON = 'battlewon',
    BATTLE_LOST = 'battlelost',
    TAKE_DAMAGE = 'damage',
    OVERLAP_PLAYER = 'overlapplayer',
    GET_POWERUP = 'getpowerup',
    BOUNCE = 'bounce',
    PLAYER_DEAD = 'playerdead',

}